requirements.txt
requests
mistralai
faiss-cpu
gradio